% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colfullnames19.R
\docType{data}
\name{colfullnames19}
\alias{colfullnames19}
\title{Countyhealthrankings.org 2019 dataset - full names of fields}
\format{
A character vector of 539 field names from "State FIPS Code" to "Population growth raw value"
}
\description{
Contains longer descriptive names of the fields than the data frame column names,
   such as "Poor or fair health raw value" instead of "v002_rawvalue"
}
\keyword{datasets}
