% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countyhealth18.R
\docType{data}
\name{countyhealth18}
\alias{countyhealth18}
\title{Countyhealthrankings.org 2018 dataset *** WORK IN PROGRESS}
\format{A data.frame with >3000 rows (Counties) and >300 columns (variables)}
\source{
2014 2015 2016 2017 2018 datasets from \url{http://www.countyhealthrankings.org/rankings/data}
  # e.g., 
  utils::browseURL('http://www.countyhealthrankings.org/rankings/data') 
  # or 
  utils::download.file(url = paste('http://www.countyhealthrankings.org/sites/default/files/', fname, sep = ''), destfile = fname)
  # maybe better to use readr::read_csv() since it is better at guessing preferred format of each column
  # require(readr)
  # countyhealth18 <- read_csv(fname)
  countyhealth18 <- utils::read.csv(fname, stringsAsFactors = FALSE)
  save(countyhealth18, file = 'countyhealth18.RData')
  
  2014/2015 data obtained 3/27/2015 and slightly modified to provide
    5 digit FIPS as character field, and ST field as copy of State field.
  The 2016 data was obtained 3/17/2016 and was not modified 
  The 2017 data was obtained 3/28/2018
  The 2018 data was obtained 3/28/2018
  Also see: \url{http://www.countyhealthrankings.org/about-project}
}
\usage{
data('countyhealth18')
}
\description{
This data set provides a variety of health indicators for each US county.
}
\keyword{datasets}
