% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/countyhealthrankings-package.R
\docType{package}
\name{countyhealthrankings-package}
\alias{countyhealthrankings}
\alias{countyhealthrankings-package}
\title{countyhealthrankings.org dataset in R}
\description{
Provides health indicators data by county, from countyhealthrankings.org, for use in R.
  Source of data: \url{http://www.countyhealthrankings.org/about-project} obtained 3/27/2015 and slightly modified to provide 5 digit FIPS as character field, and ST field as copy of State field.
}

