% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lead.zeroes.R
\name{lead.zeroes}
\alias{lead.zeroes}
\title{Add leading zeroes as needed}
\usage{
lead.zeroes(fips, length.desired)
}
\arguments{
\item{fips}{Character vector, which can be FIPS codes or other data. Required.}

\item{length.desired}{A single numeric value (recycled), or vector of numbers, required, specifying how many characters long each returned string should be.}
}
\value{
Returns a vector of same length as input parameter
}
\description{
Returns the vector that was supplied, but with leading zeroes added where needed to make all elements have specified number of characters.
}
\details{
This function can be useful in working with Census data where FIPS codes are often used.
Moving data to and from a spreadsheet can remove leading zeroes that may be necessary for proper data management.
This can apply to e.g., FIPS code for a block, block group, tract, county, or state.
Note: Number of digits in FIPS codes, assuming leading zeroes are there:\cr
state		2	(2 cumulative)\cr
county	  3	(5 cum)\cr
tract		6	(11 cum) (note 11 digits is ambiguous if not sure leading zero is there)\cr
block group 	1	(12 cum) (note 12 digits is ambiguous if not sure leading zero is there)\cr
block 	  1	(13 cum)\cr
}
\examples{
lead.zeroes(c('234','01234','3'), 5)
}
