% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/put.first.R
\name{put.first}
\alias{put.first}
\title{Simple way to put certain cols first, in a data.frame}
\usage{
put.first(x, fields)
}
\arguments{
\item{x}{Required data.frame that will have its columns reordered}

\item{fields}{required character vector of strings that are among the elements of names(x)}
}
\value{
Returns a transformed data.frame with cols in new order
}
\description{
Returns a data.frame with specified columns put first, before the others.
}
\examples{
before <- data.frame(year=c(2,2,2), ID=3, numbers=4, last=1)
put.first(before, c('ID', 'numbers'))
after <- put.first(before, names(before)[length(before)] ) # put last column first
before; after
}

