% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.chsi.R
\name{urls.chsi}
\alias{urls.chsi}
\title{URLs for Community Health Status Indicators for US Counties from CDC}
\usage{
urls.chsi(fips = "http://www.cdc.gov/CommunityHealth", type = "health",
  launch = TRUE)
}
\arguments{
\item{fips}{Single fips or a vector. Optional (defaults to homepage) character
FIPS code of County (5 characters), or "Name of County, State" format. (e.g., "Montgomery County, MD").
Attempts to replace any missing leading zero. Ability to enter County name as fips is not yet working.}

\item{type}{Optional (default is health) vector (same length as fips) specifying type of indicator or report to show for URL(s). Can be any of the following:
health, demog, pm, highways, poverty. Others may be added later.}

\item{launch}{Optional (default is TRUE but only if just one fips is specified) logical, specifying whether to launch browser to display website. Ignored if >1 fips provided.}
}
\value{
Returns character URL(s).
}
\description{
Returns URLs with public health information about US Counties,
  and can also launch a browser to open a single webpage if just 1 specified.
  Information is from \url{http://wwwn.cdc.gov/CommunityHealth}
}
\details{
*** WARNING: Not yet tested for all counties, so multi-word and and nonstandard names are unlikely to work yet.
}
\examples{
  urls.chsi('01005')
  urls.chsi('01005', 'demog')
  urls.chsi('06037', 'pm')
  urls.chsi(c('31165', 31165, 1001, 0))
}
\seealso{
\code{\link{urls.countyhealthrankings}}, \code{\link[ejanalysis]{get.county.info}} from \pkg{ejanalysis} package, and \code{\link[choroplethr]{get_county_demographics}} from \pkg{choroplethr} package.
}
